+++ IsekaiMod/Content/Features/ExceptionalFeats/ExceptionalSummoningSelection.cs
@@ -1,4 +1,5 @@
-﻿using IsekaiMod.Extensions;
+﻿using IsekaiMod.Components;
+using IsekaiMod.Extensions;
 using IsekaiMod.Utilities;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes.Prerequisites;
@@ -7,6 +8,7 @@ using Kingmaker.Designers.Mechanics.Buffs;
 using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
+using Kingmaker.Localization;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.FactLogic;
 using Kingmaker.UnitLogic.Mechanics.Components;
@@ -17,11 +19,11 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
     {
         public static void Add()
         {
-            // Exceptional Summoning
+            // Mighty Summoning
             var Icon_ExceptionalSummoning = AssetLoader.LoadInternal("Features", "ICON_EXCEPTIONAL_SUMMONING.png");
-            var ExceptionalSummoningSummonBuff = Helpers.CreateBuff("ExceptionalSummoningSummonBuff", bp => {
-                bp.SetName("Exceptional Summon");
-                bp.SetDescription("This creature gets a +10 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, AC, and saving throws per character level.");
+            var MightySummoningSummonBuff = Helpers.CreateBuff("MightySummoningSummonBuff", bp => {
+                bp.SetName("Mighty Summon");
+                bp.SetDescription("This creature gets a +5 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, and AC per character level.");
                 bp.m_Icon = Icon_ExceptionalSummoning;
                 bp.Stacking = StackingType.Replace;
                 bp.IsClassFeature = true;
@@ -41,6 +43,71 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
                     c.Stat = StatType.AC;
                     c.Value = Values.CreateContextRankValue(AbilityRankType.StatBonus);
                 });
+                bp.AddComponent<ContextRankConfig>(c => {
+                    c.m_Type = AbilityRankType.Default;
+                    c.m_BaseValueType = ContextRankBaseValueType.CharacterLevel;
+                    c.m_Progression = ContextRankProgression.MultiplyByModifier;
+                    c.m_StepLevel = 5;
+                });
+                bp.AddComponent<ContextRankConfig>(c => {
+                    c.m_Type = AbilityRankType.StatBonus;
+                    c.m_BaseValueType = ContextRankBaseValueType.CharacterLevel;
+                    c.m_Progression = ContextRankProgression.AsIs;
+                });
+            });
+            var MightySummoningBuff = Helpers.CreateBuff("MightySummoningBuff", bp => {
+                bp.SetName("Mighty Summoning");
+                bp.m_Description = new LocalizedString();
+                bp.m_Icon = Icon_ExceptionalSummoning;
+                bp.Stacking = StackingType.Replace;
+                bp.IsClassFeature = true;
+                bp.m_Flags = BlueprintBuff.Flags.StayOnDeath | BlueprintBuff.Flags.HiddenInUi;
+            });
+            var MightySummoningAbility = Helpers.CreateActivatableAbility("MightySummoningAbility", bp => {
+                bp.SetName("Mighty Summoning");
+                bp.SetDescription("Your summoned creatures get a +5 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, and AC per character level.");
+                bp.m_Icon = Icon_ExceptionalSummoning;
+                bp.m_Buff = MightySummoningBuff.ToReference<BlueprintBuffReference>();
+            });
+            var MightySummoningFeature = Helpers.CreateFeature("MightySummoningFeature", bp => {
+                bp.SetName("Mighty Summoning");
+                bp.SetDescription("Your summoned creatures get a +5 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, and AC per character level.");
+                bp.m_Icon = Icon_ExceptionalSummoning;
+                bp.AddComponent<AddFacts>(c => {
+                    c.m_Facts = new BlueprintUnitFactReference[] { MightySummoningAbility.ToReference<BlueprintUnitFactReference>() };
+                });
+                bp.ReapplyOnLevelUp = true;
+            });
+            MightySummoningBuff.AddComponent<OnSpawnBuff>(c => {
+                c.m_IfHaveFact = MightySummoningFeature.ToReference<BlueprintFeatureReference>();
+                c.m_buff = MightySummoningSummonBuff.ToReference<BlueprintBuffReference>();
+                c.IsInfinity = true;
+            });
+
+            // Magical Summoning
+            var MagicalSummoningSummonBuff = Helpers.CreateBuff("MagicalSummoningSummonBuff", bp => {
+                bp.SetName("Magical Summon");
+                bp.SetDescription("This creature gets a +5 bonus to maximum Hit Points per character level and a +1 bonus to spell penetration, spell DC, spell damage, "
+                    + "and saving throws per character level.");
+                bp.m_Icon = Icon_ExceptionalSummoning;
+                bp.Stacking = StackingType.Replace;
+                bp.IsClassFeature = true;
+                bp.AddComponent<AddContextStatBonus>(c => {
+                    c.Stat = StatType.HitPoints;
+                    c.Value = Values.CreateContextRankValue(AbilityRankType.Default);
+                });
+                bp.AddComponent<SpellPenetrationBonus>(c => {
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
+                    c.Value = Values.CreateContextRankValue(AbilityRankType.StatBonus);
+                });
+                bp.AddComponent<IncreaseAllSpellsDC>(c => {
+                    c.SpellsOnly = true;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
+                    c.Value = Values.CreateContextRankValue(AbilityRankType.StatBonus);
+                });
+                bp.AddComponent<IncreaseSpellDamage>(c => {
+                    c.DamageBonus = Values.CreateContextRankValue(AbilityRankType.StatBonus);
+                });
                 bp.AddComponent<AddContextStatBonus>(c => {
                     c.Stat = StatType.SaveFortitude;
                     c.Value = Values.CreateContextRankValue(AbilityRankType.StatBonus);
@@ -57,7 +124,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
                     c.m_Type = AbilityRankType.Default;
                     c.m_BaseValueType = ContextRankBaseValueType.CharacterLevel;
                     c.m_Progression = ContextRankProgression.MultiplyByModifier;
-                    c.m_StepLevel = 10;
+                    c.m_StepLevel = 5;
                 });
                 bp.AddComponent<ContextRankConfig>(c => {
                     c.m_Type = AbilityRankType.StatBonus;
@@ -65,32 +132,34 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
                     c.m_Progression = ContextRankProgression.AsIs;
                 });
             });
-            var ExceptionalSummoningBuff = Helpers.CreateBuff("ExceptionalSummoningBuff", bp => {
-                bp.SetName("Exceptional Summoning");
-                bp.SetDescription("Your summoned creatures a +10 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, AC, and saving throws per character level.");
+            var MagicalSummoningBuff = Helpers.CreateBuff("MagicalSummoningBuff", bp => {
+                bp.SetName("Magical Summoning");
+                bp.m_Description = new LocalizedString();
                 bp.m_Icon = Icon_ExceptionalSummoning;
                 bp.Stacking = StackingType.Replace;
                 bp.IsClassFeature = true;
                 bp.m_Flags = BlueprintBuff.Flags.StayOnDeath | BlueprintBuff.Flags.HiddenInUi;
             });
-            var ExceptionalSummoningAbility = Helpers.CreateActivatableAbility("ExceptionalSummoningAbility", bp => {
-                bp.SetName("Exceptional Summoning");
-                bp.SetDescription("Your summoned creatures get a +10 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, AC, and saving throws per character level.");
+            var MagicalSummoningAbility = Helpers.CreateActivatableAbility("MagicalSummoningAbility", bp => {
+                bp.SetName("Magical Summoning");
+                bp.SetDescription("Your summoned creatures get a +5 bonus to maximum Hit Points per character level and a +1 bonus to spell penetration, spell DC, spell damage, "
+                    + "and saving throws per character level.");
                 bp.m_Icon = Icon_ExceptionalSummoning;
-                bp.m_Buff = ExceptionalSummoningBuff.ToReference<BlueprintBuffReference>();
+                bp.m_Buff = MagicalSummoningBuff.ToReference<BlueprintBuffReference>();
             });
-            var ExceptionalSummoningFeature = Helpers.CreateFeature("ExceptionalSummoningFeature", bp => {
-                bp.SetName("Exceptional Summoning");
-                bp.SetDescription("Your summoned creatures get a +10 bonus to maximum Hit Points per character level and a +1 bonus to Attack, damage, AC, and saving throws per character level.");
+            var MagicalSummoningFeature = Helpers.CreateFeature("MagicalSummoningFeature", bp => {
+                bp.SetName("Magical Summoning");
+                bp.SetDescription("Your summoned creatures get a +5 bonus to maximum Hit Points per character level and a +1 bonus to spell penetration, spell DC, spell damage, "
+                    + "and saving throws per character level.");
                 bp.m_Icon = Icon_ExceptionalSummoning;
                 bp.AddComponent<AddFacts>(c => {
-                    c.m_Facts = new BlueprintUnitFactReference[] { ExceptionalSummoningAbility.ToReference<BlueprintUnitFactReference>() };
+                    c.m_Facts = new BlueprintUnitFactReference[] { MagicalSummoningAbility.ToReference<BlueprintUnitFactReference>() };
                 });
                 bp.ReapplyOnLevelUp = true;
             });
-            ExceptionalSummoningBuff.AddComponent<OnSpawnBuff>(c => {
-                c.m_IfHaveFact = ExceptionalSummoningFeature.ToReference<BlueprintFeatureReference>();
-                c.m_buff = ExceptionalSummoningSummonBuff.ToReference<BlueprintBuffReference>();
+            MagicalSummoningBuff.AddComponent<OnSpawnBuff>(c => {
+                c.m_IfHaveFact = MagicalSummoningFeature.ToReference<BlueprintFeatureReference>();
+                c.m_buff = MagicalSummoningSummonBuff.ToReference<BlueprintBuffReference>();
                 c.IsInfinity = true;
             });
 
@@ -144,7 +213,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
             });
             var ForbiddenSummoningBuff = Helpers.CreateBuff("ForbiddenSummoningBuff", bp => {
                 bp.SetName("Forbidden Summoning");
-                bp.SetDescription("Your summoned creatures gain a +10 bonus to hit points per character level and a +1 bonus to all attributes per character level.");
+                bp.m_Description = new LocalizedString();
                 bp.m_Icon = Icon_ForbiddenSummoning;
                 bp.Stacking = StackingType.Replace;
                 bp.IsClassFeature = true;
@@ -163,6 +232,14 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] { ForbiddenSummoningAbility.ToReference<BlueprintUnitFactReference>() };
                 });
+                bp.AddComponent<PrerequisiteFeature>(c => {
+                    c.Group = Prerequisite.GroupType.All;
+                    c.m_Feature = MightySummoningFeature.ToReference<BlueprintFeatureReference>();
+                });
+                bp.AddComponent<PrerequisiteFeature>(c => {
+                    c.Group = Prerequisite.GroupType.All;
+                    c.m_Feature = MagicalSummoningFeature.ToReference<BlueprintFeatureReference>();
+                });
                 bp.ReapplyOnLevelUp = true;
             });
             ForbiddenSummoningBuff.AddComponent<OnSpawnBuff>(c => {
@@ -175,8 +252,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
             var Icon_FerociousSummoning = AssetLoader.LoadInternal("Features", "ICON_FEROCIOUS_SUMMONING.png");
             var FerociousSummoningSummonBuff = Helpers.CreateBuff("FerociousSummoningSummonBuff", bp => {
                 bp.SetName("Ferocious Summon");
-                bp.SetDescription("This creature has 2 additional attacks and gains a +10 bonus to speed. It also gains 1d6 sneak attack per character level. "
-                    + "This effect does not stack with other effects that grant additional attack.");
+                bp.SetDescription("This creature has 2 additional attacks and gains a +10 bonus to speed. It also gains 1d6 sneak attack per character level.");
                 bp.m_Icon = Icon_FerociousSummoning;
                 bp.Stacking = StackingType.Replace;
                 bp.IsClassFeature = true;
@@ -200,8 +276,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
             });
             var FerociousSummoningBuff = Helpers.CreateBuff("FerociousSummoningBuff", bp => {
                 bp.SetName("Ferocious Summoning");
-                bp.SetDescription("Your summoned creatures have 2 additional attacks and gain a +10 bonus to speed. It also gains 1d6 sneak attack per character level. "
-                    + "This effect does not stack with other effects that grant additional attack.");
+                bp.m_Description = new LocalizedString();
                 bp.m_Icon = Icon_FerociousSummoning;
                 bp.Stacking = StackingType.Replace;
                 bp.IsClassFeature = true;
@@ -209,22 +284,16 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
             });
             var FerociousSummoningAbility = Helpers.CreateActivatableAbility("FerociousSummoningAbility", bp => {
                 bp.SetName("Ferocious Summoning");
-                bp.SetDescription("Your summoned creatures have 2 additional attacks and gain a +10 bonus to speed. It also gains 1d6 sneak attack per character level. "
-                    + "This effect does not stack with other effects that grant additional attack.");
+                bp.SetDescription("Your summoned creatures have 2 additional attacks and gain a +10 bonus to speed. It also gains 1d6 sneak attack per character level.");
                 bp.m_Icon = Icon_FerociousSummoning;
                 bp.m_Buff = FerociousSummoningBuff.ToReference<BlueprintBuffReference>();
             });
             var FerociousSummoningFeature = Helpers.CreateFeature("FerociousSummoningFeature", bp => {
                 bp.SetName("Ferocious Summoning");
-                bp.SetDescription("Your summoned creatures have 2 additional attacks and gain a +10 bonus to speed. It also gains 1d6 sneak attack per character level. "
-                    + "This effect does not stack with other effects that grant additional attack.");
+                bp.SetDescription("Your summoned creatures have 2 additional attacks and gain a +10 bonus to speed. It also gains 1d6 sneak attack per character level.");
                 bp.m_Icon = Icon_FerociousSummoning;
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] { FerociousSummoningAbility.ToReference<BlueprintUnitFactReference>() };
-                });
-                bp.AddComponent<PrerequisiteFeature>(c => {
-                    c.Group = Prerequisite.GroupType.All;
-                    c.m_Feature = ExceptionalSummoningFeature.ToReference<BlueprintFeatureReference>();
                 });
                 bp.AddComponent<PrerequisiteFeature>(c => {
                     c.Group = Prerequisite.GroupType.All;
@@ -240,7 +309,8 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats
 
             var ExceptionalSummoingFeatures = new BlueprintFeatureReference[]
             {
-                ExceptionalSummoningFeature.ToReference<BlueprintFeatureReference>(),
+                MightySummoningFeature.ToReference<BlueprintFeatureReference>(),
+                MagicalSummoningFeature.ToReference<BlueprintFeatureReference>(),
                 ForbiddenSummoningFeature.ToReference<BlueprintFeatureReference>(),
                 FerociousSummoningFeature.ToReference<BlueprintFeatureReference>(),
             };